Checkpoint 2.0.
