+++ .gitignore
@@ -1,5 +1,5 @@
 *.sublime-workspace
 .DS_Store
-build/
+dist/
 node_modules
 npm-debug.log
