+++ index.js
@@ -1,0 +1,2 @@
+export * from "./client";
+export {default as topology} from "./src/topology/index";
