--- lib/topojson/compute-id.js
@@ -1,28 +1,0 @@
-// Given a hash of GeoJSON objects and an id function, invokes the id function
-// to compute a new id for each object that is a feature. The function is passed
-// the feature and is expected to return the new feature id, or null if the
-// feature should not have an id.
-module.exports = function(objects, id) {
-  if (arguments.length < 2) id = function(d) { return d.id; };
-
-  function idObject(object) {
-    if (object && idObjectType.hasOwnProperty(object.type)) idObjectType[object.type](object);
-  }
-
-  function idFeature(feature) {
-    var i = id(feature);
-    if (i == null) delete feature.id;
-    else feature.id = i;
-  }
-
-  var idObjectType = {
-    Feature: idFeature,
-    FeatureCollection: function(collection) { collection.features.forEach(idFeature); }
-  };
-
-  for (var key in objects) {
-    idObject(objects[key]);
-  }
-
-  return objects;
-};
