--- lib/topojson/coordinate-systems.js
@@ -1,4 +1,0 @@
-module.exports = {
-  cartesian: require("./cartesian"),
-  spherical: require("./spherical")
-};
