--- lib/topojson/simplify.js
@@ -1,108 +1,0 @@
-var topojson = require("../../build/topojson"),
-    systems = require("./coordinate-systems");
-
-module.exports = function(topology, options) {
-  var minimumArea = 0,
-      retainProportion,
-      verbose = false,
-      system = null,
-      N = topology.arcs.reduce(function(p, v) { return p + v.length; }, 0),
-      M = 0;
-
-  if (options)
-    "minimum-area" in options && (minimumArea = +options["minimum-area"]),
-    "coordinate-system" in options && (system = systems[options["coordinate-system"]]),
-    "retain-proportion" in options && (retainProportion = +options["retain-proportion"]),
-    "verbose" in options && (verbose = !!options["verbose"]);
-
-  topojson.presimplify(topology, system.triangleArea);
-
-  if (retainProportion) {
-    var areas = [];
-    topology.arcs.forEach(function(arc) {
-      arc.forEach(function(point) {
-        if (isFinite(point[2])) areas.push(point[2]); // ignore endpoints
-      });
-    });
-    var n = areas.length;
-    options["minimum-area"] = minimumArea = n ? areas.sort(function(a, b) { return b - a; })[Math.max(0, Math.ceil((N - 1) * retainProportion + n - N))] : 0;
-    if (verbose) process.stderr.write("simplification: effective minimum area " + minimumArea.toPrecision(3) + "\n");
-  }
-
-  topology.arcs.forEach(topology.transform ? function(arc) {
-    var dx = 0,
-        dy = 0, // accumulate removed points
-        i = -1,
-        j = -1,
-        n = arc.length,
-        source,
-        target;
-
-    while (++i < n) {
-      source = arc[i];
-      if (source[2] >= minimumArea) {
-        target = arc[++j];
-        target[0] = source[0] + dx;
-        target[1] = source[1] + dy;
-        dx = dy = 0;
-      } else {
-        dx += source[0];
-        dy += source[1];
-      }
-    }
-
-    arc.length = ++j;
-  } : function(arc) {
-    var i = -1,
-        j = -1,
-        n = arc.length,
-        point;
-
-    while (++i < n) {
-      point = arc[i];
-      if (point[2] >= minimumArea) {
-        arc[++j] = point;
-      }
-    }
-
-    arc.length = ++j;
-  });
-
-  // Remove computed area (z) for each point, and remove coincident points.
-  // This is done as a separate pass because some coordinates may be shared
-  // between arcs (such as the last point and first point of a cut line).
-  // If the entire arc is empty, retain at least two points (per spec).
-  topology.arcs.forEach(topology.transform ? function(arc) {
-    var i = 0,
-        j = 0,
-        n = arc.length,
-        p = arc[0];
-    p.length = 2;
-    while (++i < n) {
-      p = arc[i];
-      p.length = 2;
-      if (p[0] || p[1]) arc[++j] = p;
-    }
-    M += arc.length = (j || 1) + 1;
-  } : function(arc) {
-    var i = 0,
-        j = 0,
-        n = arc.length,
-        p = arc[0],
-        x0 = p[0],
-        y0 = p[1],
-        x1,
-        y1;
-    p.length = 2;
-    while (++i < n) {
-      p = arc[i], x1 = p[0], y1 = p[1];
-      p.length = 2;
-      if (x0 !== x1 || y0 !== y1) arc[++j] = p, x0 = x1, y0 = y1;
-    }
-    M += arc.length = (j || 1) + 1;
-  });
-
-  if (verbose) process.stderr.write("simplification: retained " + M + " / " + N + " points (" + Math.round((M / N) * 100) + "%)\n");
-
-  return topology;
-};
