--- lib/topojson/topology/README.md
@@ -1,2 +1,0 @@
-For an explanation of topology construction, see
-[How To Infer Topology](http://bost.ocks.org/mike/topology/).
