+++ package.json
@@ -4,8 +4,7 @@
   "description": "An extension to GeoJSON that encodes topology.",
   "keywords": [
     "topojson",
-    "geojson",
-    "shapefile"
+    "geojson"
   ],
   "homepage": "https://github.com/mbostock/topojson",
   "license": "BSD-3-Clause",
@@ -13,37 +12,30 @@
     "name": "Mike Bostock",
     "url": "http://bost.ocks.org/mike"
   },
-  "main": "server.js",
-  "jsnext:main": "client",
-  "browser": "build/topojson.js",
+  "main": "dist/topojson-server.js",
+  "module": "index",
+  "browser": "dist/topojson.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/mbostock/topojson.git"
   },
   "bin": {
-    "topojson": "./bin/topojson",
-    "topojson-geojson": "./bin/topojson-geojson",
-    "topojson-group": "./bin/topojson-group",
-    "topojson-merge": "./bin/topojson-merge",
-    "topojson-svg": "./bin/topojson-svg"
+    "topojson": "./bin/topojson"
   },
   "scripts": {
-    "pretest": "rm -rf build && mkdir build && node -e 'process.stdout.write(\"var version = \\\"\" + require(\"./package.json\").version + \"\\\"; export * from \\\"../client\\\"; export {version};\");' > build/bundle.js && rollup -f umd -n topojson -o build/topojson.js -- build/bundle.js",
-    "test": "tape 'test/**/*-test.js' && eslint client.js src",
-    "prepublish": "npm run test && uglifyjs build/topojson.js -c -m -o build/topojson.min.js",
-    "postpublish": "VERSION=`node -e 'console.log(require(\"./package.json\").version)'`; git push && git push --tags && cp -v README.md LICENSE build/topojson.js build/topojson.min.js ../topojson-bower && cd ../topojson-bower && git add README.md LICENSE topojson.js topojson.min.js && git commit -m \"Release $VERSION.\" && git tag -am \"Release $VERSION.\" v${VERSION} && git push && git push --tags && cd ../topojson && cp build/topojson.js ../d3.github.com/topojson.v1.js && cp build/topojson.min.js ../d3.github.com/topojson.v1.min.js && cd ../d3.github.com && git add topojson.v1.js topojson.v1.min.js && git commit -m \"topojson ${VERSION}\" && git push && cd - && zip -j build/topojson.zip -- LICENSE README.md build/topojson.js build/topojson.min.js"
+    "pretest": "rm -rf dist && mkdir dist && rollup --banner \"$(preamble)\" -f umd -n topojson -o dist/topojson.js -- client.js && rollup --banner \"$(preamble)\" -f umd -n topojson -o dist/topojson-server.js -- index.js",
+    "test": "tape 'test/**/*-test.js' && eslint index.js client.js src",
+    "prepublish": "npm run test && uglifyjs dist/topojson.js -c -m -o dist/topojson.min.js && uglifyjs dist/topojson-server.js -c -m -o dist/topojson-server.min.js",
+    "postpublish": "VERSION=`node -e 'console.log(require(\"./package.json\").version)'`; git push && git push --tags && cp -v README.md LICENSE dist/topojson.js dist/topojson.min.js ../topojson-bower && cd ../topojson-bower && git add README.md LICENSE topojson.js topojson.min.js && git commit -m \"Release $VERSION.\" && git tag -am \"Release $VERSION.\" v${VERSION} && git push && git push --tags && cd ../topojson && cp dist/topojson.js ../d3.github.com/topojson.v2.js && cp dist/topojson.min.js ../d3.github.com/topojson.v2.min.js && cd ../d3.github.com && git add topojson.v2.js topojson.v2.min.js && git commit -m \"topojson ${VERSION}\" && git push && cd - && zip -j dist/topojson.zip -- LICENSE README.md dist/topojson.js dist/topojson.min.js dist/topojson-server.js dist/topojson-server.min.js"
   },
   "dependencies": {
-    "d3": "3",
-    "d3-geo-projection": "0.2",
-    "d3-queue": "2",
-    "optimist": "0.3",
-    "rw": "1",
-    "shapefile": "0.3"
+    "commander": "2",
+    "rw": "1"
   },
   "devDependencies": {
     "eslint": "3",
-    "rollup": "0.34",
+    "package-preamble": "0.0",
+    "rollup": "0.36",
     "tape": "4",
     "uglify-js": "2",
     "us-atlas": "0.0",
