+++ src/topology/README.md
@@ -1,0 +1,1 @@
+For an explanation of topology construction, see [How To Infer Topology](https://bost.ocks.org/mike/topology/).
