+++ src/topology/dedup.js
@@ -1,38 +1,38 @@
-var join = require("./join"),
-    hashmap = require("./hashmap"),
-    hashPoint = require("./point-hash"),
-    equalPoint = require("./point-equal");
+import equalPoint from "./point-equal";
+import hashmap from "./hashmap";
+import hashPoint from "./point-hash";
 
 // Given a cut topology, combines duplicate arcs.
-module.exports = function(topology) {
+export default function(topology) {
   var coordinates = topology.coordinates,
-      lines = topology.lines,
-      rings = topology.rings,
-      arcCount = lines.length + rings.length;
+      lines = topology.lines, line,
+      rings = topology.rings, ring,
+      arcCount = lines.length + rings.length,
+      i, n;
 
   delete topology.lines;
   delete topology.rings;
 
   // Count the number of (non-unique) arcs to initialize the hashmap safely.
-  for (var i = 0, n = lines.length; i < n; ++i) {
-    var line = lines[i]; while (line = line.next) ++arcCount;
+  for (i = 0, n = lines.length; i < n; ++i) {
+    line = lines[i]; while (line = line.next) ++arcCount;
   }
-  for (var i = 0, n = rings.length; i < n; ++i) {
-    var ring = rings[i]; while (ring = ring.next) ++arcCount;
+  for (i = 0, n = rings.length; i < n; ++i) {
+    ring = rings[i]; while (ring = ring.next) ++arcCount;
   }
 
   var arcsByEnd = hashmap(arcCount * 2 * 1.4, hashPoint, equalPoint),
       arcs = topology.arcs = [];
 
-  for (var i = 0, n = lines.length; i < n; ++i) {
-    var line = lines[i];
+  for (i = 0, n = lines.length; i < n; ++i) {
+    line = lines[i];
     do {
       dedupLine(line);
     } while (line = line.next);
   }
 
-  for (var i = 0, n = rings.length; i < n; ++i) {
-    var ring = rings[i];
+  for (i = 0, n = rings.length; i < n; ++i) {
+    ring = rings[i];
     if (ring.next) { // arc is no longer closed
       do {
         dedupLine(ring);
@@ -45,13 +45,14 @@ module.exports = function(topology) {
   function dedupLine(arc) {
     var startPoint,
         endPoint,
-        startArcs,
-        endArcs;
+        startArcs, startArc,
+        endArcs, endArc,
+        i, n;
 
     // Does this arc match an existing arc in order?
     if (startArcs = arcsByEnd.get(startPoint = coordinates[arc[0]])) {
-      for (var i = 0, n = startArcs.length; i < n; ++i) {
-        var startArc = startArcs[i];
+      for (i = 0, n = startArcs.length; i < n; ++i) {
+        startArc = startArcs[i];
         if (equalLine(startArc, arc)) {
           arc[0] = startArc[0];
           arc[1] = startArc[1];
@@ -62,8 +63,8 @@ module.exports = function(topology) {
 
     // Does this arc match an existing arc in reverse order?
     if (endArcs = arcsByEnd.get(endPoint = coordinates[arc[1]])) {
-      for (var i = 0, n = endArcs.length; i < n; ++i) {
-        var endArc = endArcs[i];
+      for (i = 0, n = endArcs.length; i < n; ++i) {
+        endArc = endArcs[i];
         if (reverseEqualLine(endArc, arc)) {
           arc[1] = endArc[0];
           arc[0] = endArc[1];
@@ -79,13 +80,15 @@ module.exports = function(topology) {
 
   function dedupRing(arc) {
     var endPoint,
-        endArcs;
+        endArcs,
+        endArc,
+        i, n;
 
     // Does this arc match an existing line in order, or reverse order?
     // Rings are closed, so their start point and end point is the same.
     if (endArcs = arcsByEnd.get(endPoint = coordinates[arc[0]])) {
-      for (var i = 0, n = endArcs.length; i < n; ++i) {
-        var endArc = endArcs[i];
+      for (i = 0, n = endArcs.length; i < n; ++i) {
+        endArc = endArcs[i];
         if (equalRing(endArc, arc)) {
           arc[0] = endArc[0];
           arc[1] = endArc[1];
@@ -101,8 +104,8 @@ module.exports = function(topology) {
 
     // Otherwise, does this arc match an existing ring in order, or reverse order?
     if (endArcs = arcsByEnd.get(endPoint = coordinates[arc[0] + findMinimumOffset(arc)])) {
-      for (var i = 0, n = endArcs.length; i < n; ++i) {
-        var endArc = endArcs[i];
+      for (i = 0, n = endArcs.length; i < n; ++i) {
+        endArc = endArcs[i];
         if (equalRing(endArc, arc)) {
           arc[0] = endArc[0];
           arc[1] = endArc[1];
@@ -181,4 +184,4 @@ module.exports = function(topology) {
   }
 
   return topology;
-};
+}
