+++ src/topology/point-equal.js
@@ -1,3 +1,3 @@
-module.exports = function(pointA, pointB) {
+export default function(pointA, pointB) {
   return pointA[0] === pointB[0] && pointA[1] === pointB[1];
-};
+}
