+++ src/topology/quantize.js
@@ -1,0 +1,92 @@
+export default function(objects, bbox, n) {
+  var x0 = bbox[0],
+      y0 = bbox[1],
+      x1 = bbox[2],
+      y1 = bbox[3],
+      kx = x1 - x0 ? n / (x1 - x0) : 1,
+      ky = y1 - y0 ? n / (y1 - y0) : 1;
+
+  function quantizePoint(coordinates) {
+    coordinates[0] = Math.floor((coordinates[0] - x0) * kx);
+    coordinates[1] = Math.floor((coordinates[1] - y0) * ky);
+    return coordinates;
+  }
+
+  function quantizeLine(coordinates) {
+    var i = 0,
+        j = 1,
+        n = coordinates.length,
+        pi = quantizePoint(coordinates[0]),
+        pj,
+        px = pi[0],
+        py = pi[1],
+        x,
+        y;
+
+    while (++i < n) {
+      pi = quantizePoint(coordinates[i]);
+      x = pi[0];
+      y = pi[1];
+      if (x !== px || y !== py) { // skip coincident points
+        pj = coordinates[j++];
+        pj[0] = px = x;
+        pj[1] = py = y;
+      }
+    }
+
+    coordinates.length = j;
+  }
+
+  function quantizeGeometry(o) {
+    if (o && quantizeGeometryType.hasOwnProperty(o.type)) quantizeGeometryType[o.type](o);
+  }
+
+  var quantizeGeometryType = {
+    GeometryCollection: function(o) {
+      o.geometries.forEach(quantizeGeometry);
+    },
+    Point: function(o) {
+      quantizePoint(o.coordinates);
+    },
+    MultiPoint: function(o) {
+      o.coordinates.forEach(quantizePoint);
+    },
+    LineString: function(o) {
+      var line = o.coordinates;
+      quantizeLine(line);
+      if (line.length < 2) line[1] = line[0]; // must have 2+
+    },
+    MultiLineString: function(o) {
+      for (var lines = o.coordinates, i = 0, n = lines.length; i < n; ++i) {
+        var line = lines[i];
+        quantizeLine(line);
+        if (line.length < 2) line[1] = line[0]; // must have 2+
+      }
+    },
+    Polygon: function(o) {
+      for (var rings = o.coordinates, i = 0, n = rings.length; i < n; ++i) {
+        var ring = rings[i];
+        quantizeLine(ring);
+        while (ring.length < 4) ring.push(ring[0]); // must have 4+
+      }
+    },
+    MultiPolygon: function(o) {
+      for (var polygons = o.coordinates, i = 0, n = polygons.length; i < n; ++i) {
+        for (var rings = polygons[i], j = 0, m = rings.length; j < m; ++j) {
+          var ring = rings[j];
+          quantizeLine(ring);
+          while (ring.length < 4) ring.push(ring[0]); // must have 4+
+        }
+      }
+    }
+  };
+
+  for (var key in objects) {
+    quantizeGeometry(objects[key]);
+  }
+
+  return {
+    scale: [1 / kx, 1 / ky],
+    translate: [x0 + (x1 - x0) / (2 * n), y0 + (y1 - x0) / (2 * n)]
+  };
+}
